move markdown parsing to server
