+++ dataweaver/apps/web/src/components/scopes/atlas/sync_store.ts
@@ -1,6 +1,5 @@
 'use client';
 
-import { marked } from 'marked';
 import { useEffect, useRef } from 'react';
 import type { CardEntry, CardType, QueryResult } from '~/server/types';
 import { useAtlasStore } from '~/store';
@@ -9,53 +8,6 @@ import type { AtlasContent, CardVariant } from './helpers';
 
 // --- Derivation functions ---
 // These convert raw QueryResult data from the store into AtlasContent for shapes.
-
-/** Build the variables table as an HTML string from a query result. */
-const buildTableHtml = (result: QueryResult): string => {
-  const entityDcid = result.entities[0]?.dcid ?? '';
-  const placeName = result.entities[0]?.name ?? '';
-  const intro = result.introduction ?? '';
-
-  let md = intro ? `${intro}\n\n` : '';
-  md += '| Statistical variable | Facet(s) | Rationale |\n';
-  md += '| --- | --- | --- |\n';
-
-  for (const variable of result.variables) {
-    const meta = result.metadata.find((m) => m.variableDcid === variable.dcid);
-    const firstFacet = meta?.facets[0];
-    const facetCell = firstFacet
-      ? `${firstFacet.source}<br>${firstFacet.earliestDate} – ${firstFacet.latestDate}${firstFacet.unit ? ` · ${firstFacet.unit}` : ''}`
-      : 'No data';
-
-    const encodedVar = encodeURIComponent(variable.name);
-    const encodedPlace = encodeURIComponent(placeName);
-    const link = `[${variable.name}](#fetch=${variable.dcid}&place=${entityDcid}&varName=${encodedVar}&placeName=${encodedPlace})`;
-
-    md += `| ${link} | ${facetCell} | ${variable.rationale ?? '—'} |\n`;
-  }
-
-  return marked.parse(md) as string;
-};
-
-/** Build the notes card HTML from a query result's introduction + insights. */
-const buildNotesHtml = (result: QueryResult): string => {
-  let md = '### About this data\n\n';
-  if (result.coverage) {
-    md += `${result.coverage}\n\n`;
-  }
-  if (result.introduction) {
-    md += `${result.introduction}\n\n`;
-  }
-
-  if (result.insights && result.insights.length > 0) {
-    md += '### Relevant insights\n\n';
-    for (const insight of result.insights) {
-      md += `- **${insight.title}**: ${insight.text}\n`;
-    }
-  }
-
-  return marked.parse(md) as string;
-};
 
 /** Derive AtlasContent for a loading placeholder card. */
 export const deriveLoadingContent = (title: string): AtlasContent => ({
@@ -68,7 +20,7 @@ export const deriveLoadingContent = (title: string): AtlasContent => ({
 export const deriveTableContent = (result: QueryResult): AtlasContent => ({
   variant: 'table',
   title: result.title,
-  body: buildTableHtml(result),
+  body: result.tableHtml ?? '',
   isLoading: false,
 });
 
@@ -76,7 +28,7 @@ export const deriveTableContent = (result: QueryResult): AtlasContent => ({
 export const deriveNotesContent = (result: QueryResult): AtlasContent => ({
   variant: 'text',
   title: `${result.title} • Notes`,
-  body: buildNotesHtml(result),
+  body: result.notesHtml ?? '',
   isLoading: false,
   followUp: result.followUps?.[0],
 });
